update compulsory lab 2

+++ Laborator2/compulsory/Location.java
@@ -1,18 +1,47 @@
 package compulsory;
-
+/**
+ * This class represents a location
+ * @author Popa Liliana
+ */
 public class Location {
+    /**
+     * This is the name of the location
+     */
     private String name;
+    /**
+     * This is the x coordinate of the location
+     */
     private double x;
+    /**
+     * This is the y coordinate of the location
+     */
     private double y;
+    /**
+     * This is the type of the location
+     */
     private LocationType type;
 
+    /**
+     * This is the default constructor of the class
+     */
     public Location() {
     }
 
+    /**
+     * This is the constructor of the class
+     * @param name This is the name of the location
+     */
     public Location(String name) {
         this.name = name;
     }
 
+    /**
+     * This is the constructor of the class
+     * @param name This is the name of the location
+     * @param x This is the x coordinate of the location
+     * @param y This is the y coordinate of the location
+     * @param type This is the type of the location
+     */
     public Location(String name, double x, double y, LocationType type) {
         this.name = name;
         this.x = x;
@@ -20,6 +49,10 @@ public class Location {
         this.type = type;
     }
 
+    /**
+     * This method returns the name of the location
+     * @return the name of the location
+     */
     public String getName() {
         return name;
     }
@@ -32,7 +65,7 @@ public class Location {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -40,7 +73,7 @@ public class Location {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
